requirements.txt
requirements_parser==0.2.0
